refactor: dashboard components
