+++ view/app/dashboard/components/system/disk-usage.tsx
@@ -8,7 +8,7 @@ import { DataTable, TableColumn } from '@/components/ui/data-table';
 import { TypographySmall, TypographyMuted } from '@/components/ui/typography';
 import { SystemMetricCard } from './system-metric-card';
 import { useSystemMetric } from '../../hooks/use-system-metric';
-import { formatPercentage } from './utils';
+import { formatPercentage } from '../utils/utils';
 import { DEFAULT_METRICS } from '../utils/constants';
 import { DiskUsageCardSkeletonContent } from './skeletons/disk-usage';
