+++ view/app/dashboard/components/system/load-average.tsx
@@ -7,7 +7,7 @@ import { TypographySmall, TypographyMuted } from '@/components/ui/typography';
 import { BarChartComponent } from '@/components/ui/bar-chart-component';
 import { SystemMetricCard } from './system-metric-card';
 import { useSystemMetric } from '../../hooks/use-system-metric';
-import { createLoadAverageChartData, createLoadAverageChartConfig } from './utils';
+import { createLoadAverageChartData, createLoadAverageChartConfig } from '../utils/utils';
 import { DEFAULT_METRICS, CHART_COLORS } from '../utils/constants';
 import { LoadAverageCardSkeletonContent } from './skeletons/load-average';
